+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog - [0.6.0] - yyyy-MM-dd
 
 ### New Features
+1. Add new maxflow path algorithm tool.
 
 ### Bug Fixes